+++ app/controllers/internal_controller.rb
@@ -7,4 +7,9 @@ class InternalController < ApplicationController
   # # POST /keep-me-active.js
   # def keep_me_active
   # end
+
+  # GET /autocomplete
+  def autocomplete
+    @users = current_user.associated_users.search(params[:search]).order(:full_name).limit(10)
+  end
 end